Get the LOG_FILE path from the environment by default
Usually app instances are created in multiple processes (eg. the service
serving HTTP requests and the flask shell). They should not log into the
same file(s) as this would break the logfiles: Either one of them can’t
log or they are not time-ordered as expected.

By allowing the environment to determine where to log it’s easy to
disable logging (eg. not providing the variable for flask shell) and
have different log files for different services (ie by setting the
environment variable in the services files).